Finish exercise 4.19-4.23

+++ part-4/blog-list/controllers/blogs.js
@@ -1,11 +1,11 @@
+require("dotenv").config();
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
-const User = require("../models/user");
+const middleware = require("../utils/middleware");
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", middleware.userExatractor, async (request, response) => {
   const body = request.body;
-
-  const user = await User.findById(body.userId);
+  const user = request.user;
 
   if (!body?.likes) {
     body["likes"] = 0;
@@ -35,10 +35,21 @@ blogsRouter.get("/", async (request, response) => {
   response.json(blogs);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
-});
+blogsRouter.delete(
+  "/:id",
+  middleware.userExatractor,
+  async (request, response) => {
+    const user = request.user;
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndDelete(request.params.id);
+      return response.status(204).end();
+    } else {
+      return response.status(401).json({ error: "token invalid" });
+    }
+  }
+);
 
 blogsRouter.put("/:id", async (request, response, next) => {
   const body = request.body;
